+++ CHANGES.rst
@@ -4,8 +4,8 @@ Changelog
 1.6.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
-
+- Fix: Regressions with version-annotations and stdout messages from buildout parser.
+  [loechel]
 
 1.6.1 (2017-04-07)
 ------------------